+++ packages/crypto/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @agentreceipts/crypto
+# @agent-receipts/crypto
 
 ## 0.1.0
 
@@ -9,4 +9,4 @@
 ### Patch Changes
 
 - Updated dependencies []:
-  - @agentreceipts/schema@0.2.0
+  - @agent-receipts/schema@0.2.0